+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- NativeHelperFacade utilities
+
+### Changed
+
+- Improved README documentation
+
 ## [1.1.1] - 2024-03-24
 
 ### Added